server/requirements.txt
fastapi
uvicorn
python-dotenv
mutagen
passlib[bcrypt]
python-multipart
python-jose[cryptography]